cleanup + handle update roles
