--- supabase/.gitignore
@@ -1,4 +1,0 @@
-# Supabase
-.branches
-.temp
-.env
